Change enemyConfig and add taking damage enemies and player

+++ Assets/Scripts/PlayerWeaponConfig.cs
@@ -8,11 +8,11 @@ public class PlayerWeaponConfig : ScriptableObject
     [SerializeField] GameObject projectile;
     [SerializeField] float timeBetweenShots = 0.1f;
     [SerializeField] float speedOfProjectile = 5f;
-    [SerializeField] float damage = 10f;
+    [SerializeField] int damage = 10;
 
 
     public GameObject GetPlayerProjectilePferab() { return projectile; }
     public float GetTimeBetweenShots() { return timeBetweenShots; }
     public float GetSpeedOfProjectile() { return speedOfProjectile; }
-    public float GetDamageProjectile() { return damage; }
+    public int GetDamageProjectile() { return damage; }
 }
